pyproject.toml
[tool.poetry]
name = "climate-hub"
version = "0.1.0"
description = "Control hub for AC Freedom compatible HVAC devices (CLI + Web API)"
authors = ["Giuseppe Nespolino <g.nespolino@gmail.com>"]
readme = "README.md"
packages = [{include = "climate_hub", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.9.0"
pycryptodome = "^3.19.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
mypy = "^1.8.0"
ruff = "^0.1.0"
black = "^23.12.0"
pre-commit = "^3.6.0"
types-aiofiles = "^23.2.0"

[tool.poetry.scripts]
climate = "climate_hub.cli.main:main"
climate-webapp = "climate_hub.webapp.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src/climate_hub --cov-report=term-missing --cov-report=html"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
files = ["src/climate_hub", "tests"]

[[tool.mypy.overrides]]
module = "aiohttp.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "Crypto.*"
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports
"tests/*" = ["ARG"]  # Unused arguments in tests

[tool.ruff.lint.isort]
known-first-party = ["climate_hub"]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
